# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SunoSathi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listeners:
    get:
      tags: [listeners]
      operationId: listListeners
      summary: Browse approved listeners (filter by gender, mood, online)
      parameters:
        - name: gender
          in: query
          required: false
          schema:
            type: string
            enum: [all, male, female, other]
        - name: mood
          in: query
          required: false
          schema:
            type: string
            enum: [all, sad, lonely, stress, anxiety, breakup, motivation]
        - name: onlyOnline
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: A list of listeners.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Listener'
...
  /listeners/{id}:
    get:
      tags: [listeners]
      operationId: getListenerById
      summary: Get a listener profile (with reviews)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Listener details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListenerDetail'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /listener/apply:
    post:
      tags: [listeners]
      operationId: applyAsListener
      summary: Submit a listener application (requires admin approval)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListenerApplicationBody'
      responses:
        '200':
          description: Application submitted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyListenerProfile'
  /listener/online:
    post:
      tags: [listeners]
      operationId: setOnlineStatus
      summary: Toggle the listener's online status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnlineStatusBody'
      responses:
        '200':
          description: Updated profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyListenerProfile'
...
  /listeners/{id}/reviews:
    get:
      tags: [listeners]
      operationId: getListenerReviews
      summary: Get the reviews for a listener
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reviews.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
    post:
      tags: [listeners]
      operationId: postListenerReview
      summary: Post a review for a listener
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostReviewBody'
      responses:
        '200':
          description: Review created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Listener:
      type: object
      required: [id, displayName, gender, bio, skills, photoUrl, ratingAverage, ratingCount, isOnline, lastSeenAt, pricePerMinuteChat, pricePerMinuteCall]
      properties:
        id:
          type: string
        displayName:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        bio:
          type: string
        skills:
          type: array
          items:
            type: string
        photoUrl:
          type: string
        ratingAverage:
          type: number
          format: float
        ratingCount:
          type: integer
        isOnline:
          type: boolean
        lastSeenAt:
          type: [string, 'null']
          format: date-time
        pricePerMinuteChat:
          type: integer
        pricePerMinuteCall:
          type: integer
    ListenerDetail:
      allOf:
        - $ref: '#/components/schemas/Listener'
        - type: object
          required: [reviews, totalSessions]
          properties:
            reviews:
              type: array
              items:
                $ref: '#/components/schemas/Review'
            totalSessions:
              type: integer
...
    MyListenerProfile:
      type: object
      required: [id, userId, displayName, gender, bio, skills, photoUrl, applicationStatus, isOnline, ratingAverage, ratingCount]
      properties:
        id:
          type: string
        userId:
          type: string
        displayName:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        bio:
          type: string
        skills:
          type: array
          items:
            type: string
        photoUrl:
          type: string
        applicationStatus:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: [string, 'null']
        isOnline:
          type: boolean
        lastSeenAt:
          type: [string, 'null']
          format: date-time
        ratingAverage:
          type: number
          format: float
        ratingCount:
          type: integer
    ListenerApplicationBody:
      type: object
      required: [displayName, gender, bio, skills, photoUrl]
      properties:
        displayName:
          type: string
          minLength: 2
          maxLength: 40
        gender:
          type: string
          enum: [male, female, other]
        bio:
          type: string
          minLength: 20
          maxLength: 600
        skills:
          type: array
          items:
            type: string
        photoUrl:
          type: string
        contactNumber:
          type: string
          nullable: true
        age:
          type: integer
          nullable: true
    OnlineStatusBody:
      type: object
      required: [isOnline]
      properties:
        isOnline:
          type: boolean
...
    Review:
      type: object
      required: [id, listenerId, reviewerName, rating, comment, createdAt]
      properties:
        id:
          type: string
        listenerId:
          type: string
        reviewerName:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
    PostReviewBody:
      type: object
      required: [rating, comment, sessionId]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          minLength: 3
          maxLength: 400
        sessionId:
          type: string
...